requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
bleach==6.2.0
cachalot==1.5.0
channels==4.2.0
channels_redis==4.2.1
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.1.6
django-redis==5.4.0
djangorestframework==3.15.2
drf-api-logger==1.1.16
drf-yasg==1.21.8
inflection==0.5.1
jsonpickle==4.0.1
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
sqlparse==0.5.3
tinydb==4.8.2
tinydb-smartcache==2.0.0
typing_extensions==4.12.2
uritemplate==4.1.1
webencodings==0.5.1
